search: rename query_converter.go to query/helpers.go
